backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
neo4j>=5.16.0
pandas>=2.0.0
numpy>=1.24.0
anthropic>=0.18.0
openai>=1.12.0
google-generativeai>=0.3.0
sentence-transformers>=2.3.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.0.0
google-search-results>=2.4.0
toml>=0.10.0
